backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
email-validator>=2.1.0.post1

# Database
beanie>=1.25.0
motor>=3.3.2
pymongo>=4.6.1  # Ensure compatibility with MongoDB 4.0 or higher

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Exchange Integration
ccxt>=4.2.1
websockets>=12.0
aiohttp>=3.9.1
aiodns>=3.0.0

# HTTP Clients
httpx>=0.26.0
#requests>=2.31.0
#urllib3>=2.1.0

# Data Processing & Analysis
pandas>=2.1.3
openpyxl>=3.1.2

# Caching & Rate Limiting
redis>=5.0.1

# Cron & Scheduling
apscheduler>=3.10.1

# Telegram Bot Integration
python-telegram-bot>=20.7

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
asgi-lifespan>=2.1.0
coverage>=7.3.2
mongomock>=4.1.2
pytest-mock>=3.12.0

# Utils and Helpers
python-dotenv>=1.0.0
rich>=13.7.0
loguru>=0.7.2

# Development Tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Production Tools
gunicorn>=21.2.0
supervisor>=4.2.5
pytz>=2022.7

# MongoDB version 4.0 or higher is required for compatibility with PyMongo and Motor.
